Module slack_push::message

Types related to Slack messages

Structs

Attachment
Attachments let you add more context to a message, making them more useful and effective

AttachmentActionConfirm
Protect users from destructive actions or particularly distinguished decisions by asking them to confirm their button click one more time

AttachmentField
Fields get displayed in a table-like way

Message
A message to send

SelectMenuOption
An option field

SelectMenuOptionGroup
An option group

Enums

AttachmentAction
Simplify complex workflows and empower users to take decisive action by adding interactive buttons to your messages