Enum slack_push::message::AttachmentAction

pub enum AttachmentAction {
    Button {
        text: String,
        name: Option<String>,
        value: Option<String>,
        url: Option<String>,
        style: Option<String>,
        confirm: Option<AttachmentActionConfirm>,
    },
    Select {
        text: String,
        name: Option<String>,
        value: Option<String>,
        confirm: Option<AttachmentActionConfirm>,
        options: Option<SelectMenuOption>,
        option_groups: Option<SelectMenuOptionGroup>,
        data_source: Option<String>,
        selected_options: Option<SelectMenuOption>,
        min_query_length: Option<u32>,
    },
}

Simplify complex workflows and empower users to take decisive action by adding interactive buttons to your messages

Variants

Button

Interactive button

Fields of Button

text: String

Text of the button

name: Option<String>

Provide a string to give this specific action a name. The name will be returned to your Action URL along with the message's callback_id when this action is invoked

value: Option<String>

Provide a string identifying this specific action. It will be sent to your Action URL along with the name and attachment's callback_id

url: Option<String>

URL to open for a link button

style: Option<String>

Style of the button, can be danger, primary or default

confirm: Option<AttachmentActionConfirm>

Confirmation for the button action

Select

Interactive menu

Fields of Select

text: String

Text of the menu

name: Option<String>

Provide a string to give this specific action a name. The name will be returned to your Action URL along with the message's callback_id when this action is invoked

value: Option<String>

Provide a string identifying this specific action. It will be sent to your Action URL along with the name and attachment's callback_id

confirm: Option<AttachmentActionConfirm>

Confirmation for the button action

options: Option<SelectMenuOption>

The individual options to appear in this menu

option_groups: Option<SelectMenuOptionGroup>

An alternate, semi-hierarchal way to list available options

data_source: Option<String>

Accepts static, users, channels, conversations, or external

selected_options: Option<SelectMenuOption>

The first element of this array will be set as the pre-selected option for this menu

min_query_length: Option<u32>

Only applies when data_source is set to external. If present, Slack will wait till the specified number of characters are entered before sending a request to your app's external suggestions API endpoint