Struct slack_push::message::Attachment

pub struct Attachment {
    pub fallback: Option<String>,
    pub color: Option<String>,
    pub pretext: Option<String>,
    pub author_name: Option<String>,
    pub author_link: Option<String>,
    pub author_icon: Option<String>,
    pub title: Option<String>,
    pub title_link: Option<String>,
    pub text: Option<String>,
    pub fields: Option<Vec<AttachmentField>>,
    pub image_url: Option<String>,
    pub thumb_url: Option<String>,
    pub footer: Option<String>,
    pub footer_icon: Option<String>,
    pub ts: Option<f32>,
    pub callback_id: Option<String>,
    pub actions: Option<Vec<AttachmentAction>>,
    pub attachment_type: Option<String>,
    pub mrkdwn_in: Option<Vec<String>>,
}

Attachments let you add more context to a message, making them more useful and effective

Fields
fallback: Option<String>

A plain-text summary of the attachment

color: Option<String>

Like traffic signals, color-coding messages can quickly communicate intent and help separate them from the flow of other messages in the timeline. Can either be one of good, warning, danger, or any hex color code

pretext: Option<String>

This is optional text that appears above the message attachment block

author_name: Option<String>

Small text used to display the author's name

author_link: Option<String>

A valid URL that will hyperlink the author_name

author_icon: Option<String>

A valid URL that displays a small 16x16px image to the left of the author_name text

title: Option<String>

Will be displayed as larger, bold text near the top of a message attachment

title_link: Option<String>

By passing a valid URL, the title text will be hyperlinked

text: Option<String>

This is the main text in a message attachment

fields: Option<Vec<AttachmentField>>

Array of Fields to display

image_url: Option<String>

A valid URL to an image file that will be displayed inside a message attachment

thumb_url: Option<String>

A valid URL to an image file that will be displayed as a thumbnail on the right side of a message attachment

footer: Option<String>

Some brief text to help contextualize and identify an attachment

footer_icon: Option<String>

To render a small icon beside your footer text, provide a publicly accessible URL string

ts: Option<f32>

By providing the ts field with an integer value in "epoch time", the attachment will display an additional timestamp value as part of the attachment's footer

callback_id: Option<String>

The provided string will act as a unique identifier for the collection of buttons within the attachment. It will be sent back to your message button action URL with each invoked action

actions: Option<Vec<AttachmentAction>>

AttachmentAction displayed

attachment_type: Option<String>

Even for message menus, remains the default value default

mrkdwn_in: Option<Vec<String>>

Enable markdown in a field