Struct slack_push::message::Message

pub struct Message {
    pub channel: Option<String>,
    pub text: String,
    pub as_user: Option<String>,
    pub attachments: Option<Vec<Attachment>>,
    pub icon_emoji: Option<String>,
    pub icon_url: Option<String>,
    pub link_names: Option<bool>,
    pub mrkdwn: Option<String>,
    pub parse: Option<String>,
    pub reply_broadcast: Option<String>,
    pub thread_ts: Option<String>,
    pub unfurl_links: Option<String>,
    pub unfurl_media: Option<String>,
    pub username: Option<String>,
}

A message to send

Fields
channel: Option<String>

Channel, private group, or IM channel to send message to. Can be an encoded ID, or a name

text: String

Text of the message to send

as_user: Option<String>

Pass true to post the message as the authed user, instead of as a bot

attachments: Option<Vec<Attachment>>

An array of structured attachments

icon_emoji: Option<String>

Emoji to use as the icon for this message

icon_url: Option<String>

URL to an image to use as the icon for this message

link_names: Option<bool>

Find and link channel names and usernames

mrkdwn: Option<String>

Enable Slack markup parsing

parse: Option<String>

Change how messages are treated

reply_broadcast: Option<String>

Used in conjunction with thread_ts and indicates whether reply should be made visible to everyone in the channel or conversation

thread_ts: Option<String>

Provide another message's ts value to make this message a reply

unfurl_links: Option<String>

Pass true to enable unfurling of primarily text-based content

unfurl_media: Option<String>

Pass false to disable unfurling of media content

username: Option<String>

Set your bot's user name